Changed in Widgets all floats to doubles

+++ Mapsui/Widgets/IWidget.cs
@@ -15,32 +15,32 @@ public interface IWidget
     /// <summary>
     /// Left, right or both margin depending on HorizontalAlignment
     /// </summary>
-    float MarginX { get; set; }
+    double MarginX { get; set; }
 
     /// <summary>
     /// Top, bottom or both marging depending on VerticalAlignment
     /// </summary>
-    float MarginY { get; set; }
+    double MarginY { get; set; }
 
     /// <summary>
     /// Position in X direction of left side for absolute alignment
     /// </summary>
-    float PositionX { get; set; }
+    double PositionX { get; set; }
 
     /// <summary>
     /// Position in Y direction of top side for absolute alignment
     /// </summary>
-    float PositionY { get; set; }
+    double PositionY { get; set; }
 
     /// <summary>
     /// Width of Widget
     /// </summary>
-    float Width { get; set; }
+    double Width { get; set; }
 
     /// <summary>
     /// Height of Widget
     /// </summary>
-    float Height { get; set; }
+    double Height { get; set; }
 
     /// <summary>
     /// The hit box of the widget. This needs to be updated from the widget renderer.
